debloat classification module
