+++ modules/classification/package.json
@@ -6,7 +6,6 @@
    "type": "module",
    "exports": {
       "./router/*": "./src/router/*.ts",
-      "./contracts/*": "./src/contracts/*.ts",
       "./workflows/*": "./src/workflows/*.ts",
       "./*": "./src/*.ts"
    },
@@ -26,7 +25,6 @@
       "@core/orpc": "workspace:*",
       "@core/pg-boss": "workspace:*",
       "@core/posthog": "workspace:*",
-      "@core/redis": "workspace:*",
       "@core/utils": "workspace:*",
       "@dbos-inc/dbos-sdk": "catalog:workers",
       "@dbos-inc/drizzle-datasource": "catalog:workers",
@@ -38,7 +36,6 @@
       "better-result": "catalog:validation",
       "dayjs": "catalog:ui",
       "drizzle-orm": "catalog:database",
-      "drizzle-zod": "catalog:database",
       "evlog": "catalog:logging",
       "pg-boss": "catalog:workers",
       "zod": "catalog:validation"
