--- modules/classification/src/constants.ts
@@ -1,19 +1,0 @@
-export const CLASSIFICATION_QUEUES = {
-   classify: "classify",
-   deriveKeywords: "derive-keywords",
-} as const;
-
-export type ClassificationQueueName =
-   (typeof CLASSIFICATION_QUEUES)[keyof typeof CLASSIFICATION_QUEUES];
-
-export const CLASSIFICATION_PROMPTS = {
-   classifyTransaction: "montte-classify-transaction",
-   deriveKeywords: "montte-derive-keywords",
-} as const;
-
-export type ClassificationPromptKey = keyof typeof CLASSIFICATION_PROMPTS;
-
-export const CLASSIFICATION_JOB_QUEUES = {
-   deriveKeywords: "classification:derive-keywords",
-   deriveKeywordsDeadLetter: "classification:derive-keywords:dead-letter",
-} as const;
