+++ modules/classification/src/workflows/setup.ts
@@ -1,24 +1,19 @@
-import "@modules/classification/workflows/classification-workflow";
 import { DrizzleDataSource } from "@dbos-inc/drizzle-datasource";
 import { env } from "@core/environment/worker";
-import type { PostHog, Prompts } from "@core/posthog/server";
+import type { Prompts } from "@core/posthog/server";
 import {
    createClassificationQueues,
    initClassificationWorkflowContext,
-} from "@modules/classification/workflows/context";
+} from "@modules/classification/workflows/classification-workflow";
 
 export async function setupClassificationWorkflows(deps: {
-   posthog: PostHog;
    prompts: Prompts;
    workerConcurrency: number;
 }) {
    await DrizzleDataSource.initializeDBOSSchema({
       connectionString: env.DATABASE_URL,
    });
-   initClassificationWorkflowContext({
-      posthog: deps.posthog,
-      prompts: deps.prompts,
-   });
+   initClassificationWorkflowContext(deps.prompts);
    const queues = createClassificationQueues({
       workerConcurrency: deps.workerConcurrency,
    });
